requirements.txt
Django==1.11.4
django-graphos==0.3.41
docutils==0.11
funcsigs==1.0.2
Jinja2==2.9.6
Markdown==2.3.1
MarkupSafe==1.0
matplotlib==1.4.3
mock==2.0.0
nose==1.3.7
numpy==1.10.1
olefile==0.44
pbr==3.1.1
Pillow==4.2.1
pkg-resources==0.0.0
Pygments==2.2.0
pymongo==2.5.1
pyparsing==2.0.3
python-dateutil==2.4.2
pytz==2017.2
PyVCF==0.6.7
redis==2.7.4
requests==1.2.0
six==1.10.0
Sphinx==1.2b1
tornado==3.1
